+++ test/rocprim/test_utils_types.hpp
@@ -237,6 +237,17 @@ T apply(BinaryOp binary_op, const T& a, const T& b)
 
 #define block_reduce_test_suite_type_def(name, suffix) block_reduce_test_suite_type_def_helper(name, suffix)
 
+#define block_sort_test_suite_type_def_helper(name, suffix) \
+    template<class Params> \
+    class name ## suffix : public ::testing::Test { \
+    public: \
+        using key_type = typename Params::input_type; \
+        using value_type = typename Params::output_type; \
+        static constexpr unsigned int block_size = Params::block_size; \
+    };
+
+#define block_sort_test_suite_type_def(name, suffix) block_sort_test_suite_type_def_helper(name, suffix)
+
 #define typed_test_suite_def_helper(name, suffix, params) TYPED_TEST_SUITE(name ## suffix, params)
 
 #define typed_test_suite_def(name, suffix, params) typed_test_suite_def_helper(name, suffix, params)
